+++ HISTORY.rst
@@ -8,6 +8,8 @@ Enhancements
 ^^^^^^^^^^^^
 
 - Add the ability to add dialogue middleware to dialogue managers
+- Add ability to denote a handler as only reachable via target_dialogue_state
+- Add ability to explicitly denote default handler
 
 
 3.3.0 (2018-05-10)