+++ HISTORY.rst
@@ -11,6 +11,14 @@ Enhancements
 - Add ability to denote a handler as only reachable via target_dialogue_state
 - Add ability to explicitly denote default handler
 
+3.3.1 (2018-06-20)
+------------------
+
+Bug Fixes
+~~~~~~~~~
+
+- ``mmworkbench blueprint`` has been temporarily disabled. Blueprints can manually downloaded from https://devcenter.mindmeld.com/bp/{blueprint_name}/app.tar.gz.
+
 
 3.3.0 (2018-05-10)
 ------------------